requirements.txt
# --------- pytorch --------- #
torch>=1.10.0
torchvision>=0.11.0
pytorch-lightning==1.8.1
torchmetrics==0.10.0

# --------- hydra --------- #
hydra-core==1.2.0
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.3.0.dev0
hydra-joblib-launcher==1.2.0

# --------- loggers --------- #
# wandb
# neptune-client
# mlflow
# comet-ml

# --------- others --------- #
pyrootutils     # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
#  sh            # for running bash commaneds in some tests (linux/macos only)

scanpy[leiden]>=1.9.1
statsmodels>=0.13.1
jupyterlab
glimix_core
